backend/requirements.txt
flake8==4.0.1
Django==2.2.19
djangorestframework==3.13.1
django-filter==21.1
python-dotenv==0.20.0
django-colorfield==0.8.0
django-cors-headers==3.7.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==9.3.0
djoser==2.1.0
isort==5.10.1
python-decouple==3.5
psycopg2-binary==2.9.3